+++ R/model_parameters.lm.R
@@ -18,8 +18,15 @@
 #' model_parameters(model)
 #' model_parameters(model, bootstrap = TRUE)
 #'
+#' # different p-value style in output
+#' mp <- model_parameters(model)
+#' print(mp, digits = 5)
+#' print(mp, digits = "scientific")
+#'
+#' # logistic regression model
 #' model <- glm(vs ~ wt + cyl, data = mtcars, family = "binomial")
 #' model_parameters(model)
+#'
 #' @return A data frame of indices related to the model's parameters.
 #' @export
 model_parameters.lm <- function(model, ci = .95, bootstrap = FALSE, iterations = 1000, ...) {